+++ package.json
@@ -6,6 +6,8 @@
   "dependencies": {
     "@ffmpeg-installer/ffmpeg": "^1.0.20",
     "@types/node-fetch": "^2.5.7",
+    "@types/express": "^4.17.8",
+    "@types/morgan": "^1.9.1",
     "discord.js": "^12.2.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
@@ -31,7 +33,14 @@
 
   },
   "devDependencies": {
-    "nodemon": "^2.0.4"
+    "@types/express": "^4.17.8",
+    "@types/jest": "^26.0.14",
+    "@types/morgan": "^1.9.1",
+    "@types/node": "^14.11.2",
+    "jest": "^26.4.2",
+    "nodemon": "^2.0.4",
+    "ts-jest": "^26.4.1",
+    "typescript": "^3.9.7"
   },
   "keywords": [],
   "author": "Bernabe Posadas Jr.",
